Move downstream tests into a separate file
